[addons] move AddonInfo (.cpp / .h) to its own folder
This is done to make it easier to understand after the subsequent changes and to have all of this related files in a folder.
