+++ xbmc/addons/addoninfo/AddonInfo.cpp
