+++ xbmc/addons/addoninfo/AddonInfo.h
