+++ xbmc/addons/addoninfo/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(SOURCES AddonInfo.cpp)
+
+set(HEADERS AddonInfo.h)
+
+core_add_library(addons_addoninfo)
